RPSLS game
requirements for this game were chosen to focus on logical structure
which that can be represented in sufficient way to the user. Firstly,
they game required to user interaction through the console. With this in
mind it was very clear what logical operations should be identified and
how they should behave regarding user input.