Blackjack
the game that is played in every casino on the planet. BlackJack itself
is comparing card game between a player and a dealer. The objective of
the game is to win against the dealer. The player can win by scoring 21
points on their two cards, which is called a blackjack, but only if
dealer does not have a blackjack. Another way to win is to score higher
than the dealer without exceeding 21 or let the dealer draw additional
cards until his hand exceeds 21.