Break the Sentry
The user interface consists of user panel on the left as usual and a
canvas. There is a picture of digital watch displayed on a canvas. User
cannot either use the user control panel to play the game or use the
mouse. This feature was initially added as part of the usability
testing. On top of the image text and polygons are drew. User can now
see the rate that is displayed as a text in percentages. Success and
total attempts are also displayed on a canvas just below the hit rate,
this is represented as a draw text.