Think and guess
There are two types of user interface. First one is the canvas where
text messages displayed. Second type of interface is user control panel
on the left, three of those are buttons and one is a text field.
The user interface is designed to display an output from the program on
the canvas making it more efficient. This is better interface to
interact with the user than the one where program’s responses were
printed to the console.