Pong
There is nothing in a control panel for the user. Only interface is the
actual canvas. When the game runs both teams are displayed as “Red Team”
and a “Blue Team”. The score is also appears for each player just below
the team name. The ball is also displayed on a canvas as well as left
and right paddle. Gutters are drew vertically on both sides of the
screen, however gutters cannot be interacted with as they are there as a
fixed drew lines on a canvas.