Space Glory
There is a concept of a space world in this game, narrative components
such are asteroids, missiles and the spaceship. (ERIK FAGERHOLT , 2009)
However the score, high-score and player’s life are all narrative
components. There is also an overheat meter, a new feature for this game
which would also be a concept of a game world. As overheating reach its
maximum limit another feature was added to create a blood effect during
the gameplay. Except that there is no other interaction in the game.